examples/pipelines/llm_dataset_creation/components/text_length_filter/requirements.txt
pyarrow>=7.0
fasttext-wheel==0.9.2
s3fs